# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Policy Bot API
  version: 1.0.0
  description: |
    RAG-based policy assistant API for Caribbean government staff.

    ## Authentication
    All endpoints except `/branding` require authentication via NextAuth session cookies.

    ## User Roles
    - **admin**: Full system access
    - **superuser**: Manage documents/users in assigned categories
    - **user**: Query documents in subscribed categories
  contact:
    name: Policy Bot Support
  license:
    name: Proprietary
servers:
  - url: https://policybot.abhirup.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
    Thread:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "HR Policy Questions"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        uploadCount:
          type: integer
          example: 1
        categoryIds:
          type: array
          items:
            type: integer
          example: [1, 3]
      required:
        - id
        - title
        - createdAt
        - updatedAt
        - uploadCount
        - categoryIds
    Message:
      type: object
      properties:
        id:
          type: string
          example: "msg_abc123"
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        attachments:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - timestamp
    Source:
      type: object
      properties:
        documentName:
          type: string
          example: "HR_Handbook_2024.pdf"
        pageNumber:
          type: integer
          example: 12
        chunkText:
          type: string
          example: "Annual leave entitlement is 20 working days..."
        score:
          type: number
          format: float
          example: 0.89
        isWeb:
          type: boolean
          description: True if source is from Tavily web search
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - ADMIN_REQUIRED
            - SUPERUSER_REQUIRED
            - NOT_FOUND
            - VALIDATION_ERROR
            - FILE_TOO_LARGE
            - UPLOAD_LIMIT
            - INVALID_FILE_TYPE
            - DUPLICATE
            - SERVICE_ERROR
        details:
          type: string
          description: Additional context
      required:
        - error
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Unauthorized"
            code: "AUTH_REQUIRED"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
...
security:
  - sessionAuth: []
paths:
...
  /threads/{threadId}:
    parameters:
      - name: threadId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Threads
      summary: Get thread
      description: Get a specific thread with all messages
      operationId: getThread
      responses:
        '200':
          description: Thread details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Thread'
                  - type: object
                    properties:
                      messages:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
                      uploads:
                        type: array
                        items:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Threads
      summary: Update thread
      description: Update thread title or categories
      operationId: updateThread
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 100
                categoryIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Thread updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  categoryIds:
                    type: array
                    items:
                      type: integer
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Threads
      summary: Delete thread
      description: Delete a thread and all associated data
      operationId: deleteThread
      responses:
        '200':
          description: Thread deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: object
                    properties:
                      threadId:
                        type: string
                      messageCount:
                        type: integer
                      uploadCount:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /threads/{threadId}/upload:
    parameters:
      - name: threadId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags:
        - Threads
      summary: Upload file to thread
      description: Upload a PDF to a thread for compliance checking
      operationId: uploadToThread
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file (max 5MB)
              required:
                - file
      responses:
        '200':
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  size:
                    type: integer
                  uploadCount:
                    type: integer
        '400':
          description: Validation error or upload limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...